% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_bg.R
\name{descale_w_objects}
\alias{descale_w_objects}
\title{Return scaled variables to the original scale using means and SDs}
\usage{
descale_w_objects(data, mean_vector, sd_vector)
}
\description{
A little function to rescale data using vectors of means and sds
}
\author{
Brian Maitner
}
