% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_CVmaxnet.R
\name{dr_CVmaxnet}
\alias{dr_CVmaxnet}
\title{Density-ratio SDM estimation with CV MAXNET}
\usage{
dr_CVmaxnet(
  presence_data = NULL,
  background_data = NULL,
  projection_data = NULL,
  f = NULL,
  regmult = 1,
  regfun = maxnet.default.regularization,
  addsamplestobackground = T,
  clamp = T,
  verbose = FALSE,
  method,
  type = c("link", "exponential", "cloglog", "logistic"),
  object = NULL
)
}
\arguments{
\item{presence_data}{dataframe of covariates}

\item{background_data}{dataframe of covariates}

\item{projection_data}{dataframe of covariates}

\item{f}{MAXNET formula to use. Default (NULL) will use the MAXNET default.}

\item{regmult}{MAXNET regularization multiplier. Default is 1.}

\item{regfun}{MAXNET regularization function. Default is the MAXNET default.}

\item{addsamplestobackground}{If TRUE (the default), any presences that aren't in the background will be added.}

\item{clamp}{If TRUE (the default), predictions will be limited to ranges seen in the training dataset.}

\item{method}{one of either "fit" or "predict"}

\item{type}{Type of response required.  Defaults to link, exponential, cloglog, and logistic.}

\item{object}{fitted object returned by a dr_... function. Only needed when method = "predict"}
}
\description{
dr_CVmaxnet is an internal function for density-ratio estimation with MAXNET
}
\note{
The options f, regmult, regfun, and addSamplestobackground are only used when method == "predict",
the options clamp and type are only used when method == "predict". See the much better documentation for maxnet for more details.
}
\keyword{internal}
