% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_ulsif.R
\name{dr_ulsif}
\alias{dr_ulsif}
\title{Density-ratio SDM estimation with uLSIF}
\usage{
dr_ulsif(
  presence_data = NULL,
  background_data = NULL,
  projection_data = NULL,
  sigma = 10^seq(-3, 1, length.out = 9),
  lambda = 10^seq(-3, 1, length.out = 9),
  kernel_num = 100,
  verbose = FALSE,
  method,
  object = NULL
)
}
\arguments{
\item{presence_data}{dataframe of covariates}

\item{background_data}{dataframe of covariates}

\item{projection_data}{dataframe of covariates}

\item{sigma}{Sigma parameter for uLSIF. Default is the uLSIF default.}

\item{lambda}{Lambda parameter for uLSIF. Default is the uLSIF default.}

\item{kernel_num}{kernel_number for uLSIF. Default is the uLSIF default.}

\item{method}{one of either "fit" or "predict"}

\item{object}{fitted object returned by a dr_... function. Only needed when method = "predict"}
}
\description{
dr_ulsif is an internal function for density-ratio estimation with uLSIF
}
\keyword{internal}
