% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_range_map.R
\name{evaluate_range_map}
\alias{evaluate_range_map}
\title{Evalute PBSDM range map quality}
\usage{
evaluate_range_map(
  occurrences,
  env,
  method = NULL,
  presence_method = NULL,
  background_method = NULL,
  bootstrap = "none",
  bootstrap_reps = 100,
  quantile = 0.05,
  constraint_regions = NULL
)
}
\arguments{
\item{occurrences}{Presence coordinates in long,lat format.}

\item{env}{Environmental rasters}

\item{method}{Optional. If supplied, both presence and background density estimation will use this method.}

\item{presence_method}{Optional. Method for estimation of presence density.}

\item{background_method}{Optional. Method for estimation of background density.}

\item{bootstrap}{Character.  One of "none" (the default, no bootstrapping),
"numbag" (presence function is bootstrapped),
or "doublebag" (presence and background functions are bootstrapped).}

\item{bootstrap_reps}{Integer.  Number of bootstrap replicates to use (default is 100)}

\item{quantile}{Quantile to use for thresholding.  Default is 0.05 (5 pct training presence). Set to 0 for minimum trainin presence (MTP).}

\item{constraint_regions}{See get_env_bg documentation}

\item{...}{Additional parameters passed to internal functions.}
}
\description{
This function uses 5-fold, spatially stratified, cross-validation to evaluate distributin model quality.
}
\details{
Current plug-and-play methods include: "gaussian", "kde","vine","rangebagging", "lobagoc", and "none".
Current density ratio methods include: "ulsif", "rulsif".
}
\note{
Either \code{method} or both \code{presence_method} and \code{background_method} must be supplied.
}
