% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_density_ratio.R
\name{fit_density_ratio}
\alias{fit_density_ratio}
\title{Fit density-ratio distribution models in a plug-and-play framework.}
\usage{
fit_density_ratio(presence = NULL, background = NULL, method = NULL, ...)
}
\arguments{
\item{presence}{dataframe of covariates at presence points}

\item{background}{Dataframe of covariates at background points}

\item{method}{Character. See "notes" for options.}

\item{...}{Additional parameters passed to internal functions.}
}
\value{
List of class "dr_model" containing model objects and metadata needed for projecting the fitted models.
}
\description{
This function fits density-ratio species distribution models for the specified density-ratio method.
}
\details{
Current methods include: "ulsif", "rulsif", "maxnet"
}
