% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_bg.R
\name{get_env_bg}
\alias{get_env_bg}
\title{Extract background data for SDM fitting.}
\usage{
get_env_bg(
  coords,
  env,
  method = "buffer",
  width = NULL,
  constraint_regions = NULL
)
}
\arguments{
\item{coords}{Coordinates (long,lat) to extract values for}

\item{env}{Environmental rasterstack in any projection}

\item{method}{Methods for getting bg points. Current option is buffer}

\item{width}{Numeric or NULL.  Width (meters or map units) of buffer. If NULL, uses max dist between nearest occurrences.}

\item{constraint_regions}{An optional spatialpolygons* object that can be used to limit the selection of background points.}
}
\value{
A list containing 1) the background data, 2) the cell indices for which the background was taken
}
\description{
This function extracts background data around known presence records.
}
\note{
If supplying constraint_regions, any polygons in which the occurrences fall are considered fair game for background selection.
This background selection is, however, still limited by the buffer as well.
}
\examples{
{

# load packages
 library(geodata)

# make temp directory

 temp <- tempdir()

# Get some occurrence data

occurrences <- BIEN::BIEN_occurrence_species(species = "Trillium vaseyi",
                                             new.world = TRUE,
                                             cultivated = FALSE)

# Thin down to unique occurrences

occurrences <- unique(occurrences[c("longitude","latitude")])

# Get bioclim data

env <- worldclim_global(var = "bio",
                        res = 10,
                        path = temp)


env <- env[[c(1,12)]]

range <- BIEN::BIEN_ranges_load_species(species = "Xanthium strumarium")


bg_data <- get_env_bg(coords = occurrences,
                      env = env,
                      method = "buffer",width = 100,
                      constraint_regions = range)


}
}
