% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_functions}
\alias{get_functions}
\title{Internal function for getting available function names.}
\usage{
get_functions(type = "pnp")
}
\arguments{
\item{type}{Type of function to get. Options are "pnp" for presence/background functions and "dr" for ratio functions.}
}
\description{
This function checks the available functions in the package to extract current options for dr and pnp fitting
}
\keyword{internal}
