% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_curves.R
\name{get_response_curves}
\alias{get_response_curves}
\title{Generate Response Curves}
\usage{
get_response_curves(
  env_bg,
  env_pres,
  pnp_model,
  n.int = 200,
  envMeans = NULL,
  envSDs = NULL
)
}
\arguments{
\item{envMeans}{A vector of means for each environmental predictor in the dataset.}

\item{envSDs}{A vector of standard deviations for each environmental predictor in the dataset.}

\item{dirs}{A list of directories that includes the algorithm directory and diagnostics directory.}

\item{env}{An \code{RasterStack} or \code{RasterBrick} of environmental predictors.}

\item{stats}{A data.frame that contains information about each model that was fit, including model names, species names, and which algorithm was used.}

\item{bg}{An \code{OccurrenceData} object containing the background data for each species.}

\item{toDo}{A list of parameters that specifies the details of the analysis to be run.}

\item{best.var}{A vector of character strings that specifies the environmental predictors to include in the response curves.}

\item{priors}{A list that contains prior distributions for each parameter in the model.}

\item{pres}{An \code{OccurrenceData} object containing the presence data for each species.}

\item{extrapLimits}{A list of limits to use for extrapolation of the response curves.}

\item{openFigs}{A logical value indicating whether to open the generated plots.}
}
\value{
This function generates a set of PDF files in the diagnostics directory for each species, with one file per model. The PDF files contain plots of the response curves for each environmental predictor in the model.
}
\description{
Given an environmental data set, fitted models, and a directory to output plots, this function generates response curves for each predictor in the model. The response curves depict the predicted change in probability of presence as a function of the environmental predictor while holding all other predictors constant at their mean values.
}
\author{
Cory Merow, modified by Brian Maitner
}
