% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnp_gaussian.R
\name{pnp_gaussian}
\alias{pnp_gaussian}
\title{Internal function for fitting gaussian distributions in plug-and-play SDMs.}
\usage{
pnp_gaussian(data, method, type = "regularized", object = NULL)
}
\arguments{
\item{data}{dataframe of covariates}

\item{method}{one of either "fit" or "predict"}

\item{type}{one of either "classical", "robust", or "regularized" (the default)}

\item{object}{fitted object returned by a pnp_... function. Only needed when method = "predict"}
}
\description{
This function both fits distributions and projects those distributions to new covariates..
}
\keyword{internal}
