% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnp_kde.R
\name{pnp_kde}
\alias{pnp_kde}
\title{Internal function for fitting KDE distributions in plug-and-play SDMs.}
\usage{
pnp_kde(data, method, object = NULL)
}
\arguments{
\item{data}{dataframe of covariates}

\item{method}{one of either "fit" or "predict"}

\item{object}{fitted object returned by a pnp_... function. Only needed when method = "predict"}
}
\description{
This function both fits Kernel Density Estimation (KDE) distributions and projects those distributions to new covariates..
}
\keyword{internal}
