% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnp_none.R
\name{pnp_none}
\alias{pnp_none}
\title{Internal function for returning empty pnp_estimate class models}
\usage{
pnp_none(method, object = NULL, ...)
}
\arguments{
\item{method}{one of either "fit" or "predict"}

\item{object}{fitted object returned by a pnp_... function. Only needed when method = "predict"}
}
\description{
This function is used internally to transform presence-background models into presence-only models.
}
\keyword{internal}
