% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnp_vine.R
\name{pnp_vine}
\alias{pnp_vine}
\title{Internal function for fitting vine copula distributions in plug-and-play SDMs.}
\usage{
pnp_vine(data, method, object = NULL)
}
\arguments{
\item{data}{dataframe of covariates}

\item{method}{one of either "fit" or "predict"}

\item{object}{fitted object returned by a pnp_... function. Only needed when method = "predict"}
}
\description{
This function both fits distributions and projects those distributions to new covariates.
}
\keyword{internal}
