% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_density_ratio.R
\name{project_density_ratio}
\alias{project_density_ratio}
\title{Projects fitted density-ratio distribution models onto new covariates.}
\usage{
project_density_ratio(dr_model, data)
}
\arguments{
\item{dr_model}{A fitted density ratio model produced by \code{fit_density_ratio}}

\item{data}{covariate data}
}
\value{
A vector of relative suitabilities evaluates at the covariates supplied in the data object.
}
\description{
This function projects fitted density-ratio species distribution models onto new covariates.
}
