% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_bg.R
\name{rescale_w_objects}
\alias{rescale_w_objects}
\title{Rescale a dataset using vectors of means and SDs}
\usage{
rescale_w_objects(data, mean_vector, sd_vector)
}
\description{
A little function to rescale data using vectors of means and sds
}
\author{
Brian Maitner
}
