% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_points}
\alias{sample_points}
\title{Example S4DM occurrence data}
\format{
A data.frame with 65 observations of 2 variables:
\describe{
\item{Longitude}{Longitude, in decimal degrees}
\item{Latitude}{Latitude, in decimal degrees}
...
}
}
\source{
\url{https://biendata.org}
}
\usage{
sample_points
}
\description{
A sample dataset containing occurrence records.
}
\keyword{datasets}
