% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_threshold.R
\name{sdm_threshold}
\alias{sdm_threshold}
\title{Thresholds a continuous "suitability" raster to create a binary raster.}
\usage{
sdm_threshold(
  prediction_raster,
  occurrence_sf,
  quantile = 0.05,
  return_binary = TRUE
)
}
\arguments{
\item{prediction_raster}{Raster containing continuous predictions of "suitability" to be thresholded}

\item{occurrence_sf}{A sf object containing presence locations. In the projection of the prediciton raster}

\item{quantile}{Numeric between 0 and 1. Quantile to use for thresholding (defaults to 0.05).  Set to 0 for minimum training presence.}

\item{return_binary}{LOGICAL. Should the raster returned be binary (presence/absence)?  If FALSE, predicted presenced will retain their 'suitability" scores.}
}
\description{
This function thresholds a continuous "suitability" raster to produce a binary presence/absence raster.
}
\author{
Cecina Babich Morrow (modified by Brian Maitner)
}
