% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_pres.R
\name{get_env_pres}
\alias{get_env_pres}
\title{Extract presence data for SDM fitting.}
\usage{
get_env_pres(coords, env)
}
\arguments{
\item{coords}{Coordinates (long,lat) to extract values for}

\item{env}{Environmental rasterstack in any projection}
}
\description{
This function extracts presence data at known presence records.
}
