% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_pres.R
\name{get_env_pres}
\alias{get_env_pres}
\title{Extract presence data for SDM fitting.}
\usage{
get_env_pres(coords, env)
}
\arguments{
\item{coords}{Coordinates (long,lat) to extract values for}

\item{env}{Environmental rasterstack in any projection}
}
\description{
This function extracts presence data at known presence records.
}
\examples{
 {

# load packages
 library(geodata)

# make temp directory

 temp <- tempdir()

# Get some occurrence data

occurrences <- BIEN::BIEN_occurrence_species(species = "Trillium vaseyi",
                                             new.world = TRUE,
                                             cultivated = FALSE)

# Thin down to unique occurrences

occurrences <- unique(occurrences[c("longitude","latitude")])

# Get bioclim data

env <- worldclim_global(var = "bio",
                        res = 10,
                        path = temp)


env <- env[[c(1,12)]]


env_pres <- get_env_pres(coords = occurrences,
                        env = env)

}
}
