% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_random.R
\name{stratify_random}
\alias{stratify_random}
\title{Split data for k-fold spatially stratified cross validation}
\usage{
stratify_random(occurrence_sp, nfolds = NULL)
}
\arguments{
\item{occurrence_sp}{a SpatialPoints or SpatialPointsDataFrame object}

\item{nfolds}{number of desired output folds.}
}
\value{
Returns a SpatialPoints dataframe with the data.frame containing fold designation for each point.
}
\description{
Splitting tool for cross-validation
}
\details{
See Examples.
}
\author{
Cory Merow \href{mailto:cory.merow@gmail.com}{cory.merow@gmail.com}
}
